configure.ac: libdrm/drm_fourcc.h detection
To compile the library with recent features, the included
header needs to meet the given kernel version and must hold
the needed defines.

This patch extends configure.ac to checks if the header
is available in libdrm's system include path.
Next it will check it's recent staga and adapt the corresponding
CFLAGS. If it can't succeed, it will fall back to kernel headers.

Signed-off-by: Ralf Zerres <ralf.zerres@networkx.de>