surface: set output format before capture format
In RequestCreateSurfaces2, the S_FMT(CAP) may not set the desired format
if the capture format is limited to the output format dimensions, unless
the output format is set in advance.

Use V4L2_PIX_FMT_H264_SLICE because we know that requires larger capture
buffers to store motion vectors on Hantro G1.

Signed-off-by: Philipp Zabel <p.zabel@pengutronix.de>